% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAdataViz.R
\name{getDataAnalytics}
\alias{getDataAnalytics}
\title{TCGA metadata statistical significance tests (regression and t-test) to compare patient variables in cancer groups}
\usage{
getDataAnalytics(
  metadata,
  x_variable_name,
  y_variable_name,
  analysis_name = "regression"
)
}
\arguments{
\item{metadata}{as filtered or unfiltered tibble from getMetaData() or from getDataViz()}

\item{x_variable_name}{independent variable of the model (continuous or categorical)}

\item{y_variable_name}{dependent variable of the model (continuous or categorical)}

\item{analysis_name}{name of analysis (linear regression or a t-test)}
}
\description{
TCGA metadata statistical significance tests (regression and t-test) to compare patient variables in cancer groups
}
