% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAdataViz.R
\name{getDataViz}
\alias{getDataViz}
\title{TCGA metadata filtering and visualization by a boxplot binned by a cancer type.
Filtering is done based on patient ids or cancer typer}
\usage{
getDataViz(
  metadata,
  cancer_type = NULL,
  patients = c(),
  metadata_col = NULL,
  metadata_levels = c(),
  phenotype = NULL
)
}
\arguments{
\item{metadata}{metadata tibble downladed from getMetaData()}

\item{cancer_type}{a list of cancer types to filter metadata on}

\item{patients}{list of TCGA patient ids in TCGA-XX-XXX format to filter metadata on}

\item{metadata_col}{optional name of the column as an extra dimension for comparison}

\item{metadata_levels}{optional list of categories to filter metadata_col on}

\item{phenotype}{name of the phenotype column to extract data from (the y axis of the boxplot)}
}
\value{
filtered_meta
}
\description{
TCGA metadata filtering and visualization by a boxplot binned by a cancer type.
Filtering is done based on patient ids or cancer typer
}
