% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAdataViz.R
\name{getMetaData}
\alias{getMetaData}
\title{A function to download patient metadata from TCGA API and filter patients based on tumor_status field}
\usage{
getMetaData(
  url = "https://api.gdc.cancer.gov/data/1b5f413e-a8d1-4d10-92eb-7c4ae739ed81"
)
}
\arguments{
\item{url}{as url to the data provided by the TCGA project}
}
\value{
returns raw metadata downlaoded from TCGA resource
}
\description{
A function to download patient metadata from TCGA API and filter patients based on tumor_status field
}
